messing around

+++ process_kitti.py
@@ -37,8 +37,12 @@ def download_data():
         if not os.path.exists(c_dir): os.mkdir(c_dir)
         for i, d in enumerate(drive_list):
             print str(i+1) + '/' + str(len(drive_list)) + ": " + d
-            url = "http://kitti.is.tue.mpg.de/kitti/raw_data/" + d + "/" + d + "_sync.zip"
-            urllib.urlretrieve(url, filename=c_dir + d + "_sync.zip")
+            if os.path.exists(c_dir + d + '_sync.zip'):
+                print '  skipping ' + c_dir + d + '_sync.zip' + ', already exists'
+            else:
+                print '  downloading to ' + c_dir + d + '_sync.zip'
+                url = "http://kitti.is.tue.mpg.de/kitti/raw_data/" + d + "/" + d + "_sync.zip"
+                urllib.urlretrieve(url, filename=c_dir + d + "_sync.zip")
 
 
 # unzip images
